feat: create zip-bundle.ts in `zipper` and export it in /lib/index.ts
